Refactor core into submodules
